refactor(ResultStatus) : bettingAmount를 받아 계산한 후 Profit을 반환하도록 변경
