+++ src/main/java/blackjack/domain/player/Players.java
@@ -1,7 +1,10 @@
 package blackjack.domain.player;
 
+import blackjack.domain.result.Result;
 import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
 public class Players {
     private final Dealer dealer;
@@ -10,6 +13,21 @@ public class Players {
     public Players(Dealer dealer, List<GamePlayer> gamePlayers) {
         this.dealer = dealer;
         this.gamePlayers = gamePlayers;
+    }
+
+    public Result compareResults() {
+        Map<Name, Profit> playerResults = new LinkedHashMap<>();
+        Profit dealerProfit = new Profit(0);
+        playerResults.put(dealer.getName(), dealerProfit);
+
+        for (GamePlayer gamePlayer : gamePlayers) {
+            Profit playerProfit = gamePlayer.getProfit(dealer);
+            playerResults.put(gamePlayer.getName(), playerProfit);
+            dealerProfit = dealerProfit.subtract(playerProfit);
+        }
+        playerResults.put(dealer.getName(), dealerProfit);
+
+        return new Result(playerResults);
     }
 
     public Dealer getDealer() {
